requirements.txt
ovos-plugin-manager>=0.0.1a2
lingua-language-detector~=1.0.1
